refactor(validators): separate artifact contracts
